feat: move ApiRequestDefinition to lib
