+++ frontend/lib/components/ApiRequestDefinition/components/ParamsDefinition/utils/mapParameterDefinitionToApiBrainDefinitionSchema.ts
